+++ blade-graphics/src/lib.rs
@@ -104,6 +104,8 @@ pub struct ContextDesc {
     pub capture: bool,
     /// Enable GAPI overlay.
     pub overlay: bool,
+    /// Force selection of a specific Device ID, unless 0.
+    pub device_id: u32,
 }
 
 #[derive(Debug)]